extract IgniteTupleCommon
